// src/interfaces/realtime.ts
export interface IRealtime {
  DC_NOME: string;
  PRE_INS: string;
  VL_LATITUDE: string;
  PRE_MAX: string;
  UF: string;
  RAD_GLO: string;
  PTO_INS: string;
  TEM_MIN: string;
  VL_LONGITUDE: string;
  UMD_MIN: string;
  PTO_MAX: string;
  VEN_DIR: string;
  DT_MEDICAO: string;
  CHUVA: string;
  PRE_MIN: string;
  UMD_MAX: string;
  VEN_VEL: string;
  PTO_MIN: string;
  TEM_MAX: string;
  TEN_BAT: string;
  VEN_RAJ: string;
  TEM_CPU: string;
  TEM_INS: string;
  UMD_INS: string;
  CD_ESTACAO: string;
  HR_MEDICAO: string;
}